Fix X11 WaitEventsIterator busy loop
Certain events were not being checked for in the PollEventsIterator
and would be left in the X11 queue forever. XPeekEvent in the
WaitEventsIterator would never block because an item was still in the
queue causing a busy loop. The problem is solved by using XNextEvent
in PollEventsIterator with a check to make sure it will never block.