Fix issue with dead keys
Glutin had two issues while handling keyboard events. The first one was
caused by passing KeyRelease events to Xutf8LookupString, which is wrong
according to the docs [1].

On the second issue Glutin was calling XFilterEvent but not using its
return. This function should advice client code to skip a event if it
was already handled [2]. This is essential for dead keys to work, since
the dead key itself should neither print nor do anything at all. To fix
this issue we now save XFilterEvent return and, if it is true, exit the
function early.

[1] https://www.x.org/archive/current/doc/man/man3/XmbLookupString.3.xhtml
[2] https://www.x.org/archive/current/doc/man/man3/XFilterEvent.3.xhtml